backend/requirements.txt
gunicorn==20.0.0
django==2.2.7
djangorestframework==3.10.3
pyrebase==3.0.27
pillow==6.2.1
firebase-admin==3.2.0
xhtml2pdf==0.2.3